+++ frontend/packages/core/src/Utils/index.tsx
@@ -1,3 +1,4 @@
+export { Popper, PopperItem } from "./popper";
 export { default as styled } from "./styled";
 export { css as EMOTION_CSS, keyframes as EMOTION_KEYFRAMES } from "@emotion/react";
