+++ frontend/packages/core/src/index.tsx
@@ -33,7 +33,6 @@ export { client } from "./Network";
 export * from "./NPS";
 export { default as ExpansionPanel } from "./panel";
 export { default as Paper } from "./paper";
-export { Popper, PopperItem } from "./popper";
 export { default as Resolver } from "./Resolver";
 export { Step, Stepper } from "./stepper";
 export * from "./Utils";
